Refactor task grouping into shared service
