+++ src/ui/views/browser/widgets/todoWidget.js
@@ -1,18 +1,12 @@
 import { formatHours } from '../../../../core/helpers.js';
-import { addLog } from '../../../../core/log.js';
-import { getState } from '../../../../core/state.js';
-import { endDay, checkDayEnd } from '../../../../game/lifecycle.js';
+import { endDay } from '../../../../game/lifecycle.js';
 import { normalizeActionEntries } from '../../../actions/registry.js';
 import {
-  DEFAULT_FOCUS_BUCKET,
-  getFocusBucketComparator,
-  getFocusModeConfig,
-  registerFocusBucket
-} from '../../../actions/focusBuckets.js';
-import { advanceActionInstance, completeActionInstance } from '../../../../game/actions/progress.js';
-import { getActionDefinition } from '../../../../game/registryService.js';
-import { spendTime } from '../../../../game/time.js';
-import { allocateDailyStudy } from '../../../../game/requirements.js';
+  applyFocusOrdering,
+  attachProgressHandlers,
+  createProgressHandler,
+  DEFAULT_TODO_EMPTY_MESSAGE
+} from '../../../actions/taskGrouping.js';
 import todoDom from './todoDom.js';
 import todoState from './todoState.js';
 
@@ -72,318 +66,6 @@ function getAvailableMoney(model = {}) {
   return Math.max(0, available);
 }
 
-function sortHustleEntries(entries = []) {
-  return [...entries].sort((a, b) => {
-    const roiA = Number.isFinite(a?.moneyPerHour) ? a.moneyPerHour : -Infinity;
-    const roiB = Number.isFinite(b?.moneyPerHour) ? b.moneyPerHour : -Infinity;
-    if (roiA !== roiB) {
-      return roiB - roiA;
-    }
-    const payoutA = Number.isFinite(a?.payout) ? a.payout : 0;
-    const payoutB = Number.isFinite(b?.payout) ? b.payout : 0;
-    if (payoutA !== payoutB) {
-      return payoutB - payoutA;
-    }
-    const durationA = Number.isFinite(a?.durationHours) ? a.durationHours : Infinity;
-    const durationB = Number.isFinite(b?.durationHours) ? b.durationHours : Infinity;
-    if (durationA !== durationB) {
-      return durationA - durationB;
-    }
-    const orderA = Number.isFinite(a?.orderIndex) ? a.orderIndex : 0;
-    const orderB = Number.isFinite(b?.orderIndex) ? b.orderIndex : 0;
-    return orderA - orderB;
-  });
-}
-
-function sortUpgradeEntries(entries = []) {
-  return [...entries].sort((a, b) => {
-    const remainingA = Number.isFinite(a?.upgradeRemaining) ? a.upgradeRemaining : Infinity;
-    const remainingB = Number.isFinite(b?.upgradeRemaining) ? b.upgradeRemaining : Infinity;
-    if (remainingA !== remainingB) {
-      return remainingA - remainingB;
-    }
-    const durationA = Number.isFinite(a?.durationHours) ? a.durationHours : Infinity;
-    const durationB = Number.isFinite(b?.durationHours) ? b.durationHours : Infinity;
-    if (durationA !== durationB) {
-      return durationA - durationB;
-    }
-    const orderA = Number.isFinite(a?.orderIndex) ? a.orderIndex : 0;
-    const orderB = Number.isFinite(b?.orderIndex) ? b.orderIndex : 0;
-    return orderA - orderB;
-  });
-}
-
-function sortCommitmentEntries(entries = []) {
-  return [...entries].sort((a, b) => {
-    const remainingA = Number.isFinite(a?.progress?.remainingDays)
-      ? a.progress.remainingDays
-      : Infinity;
-    const remainingB = Number.isFinite(b?.progress?.remainingDays)
-      ? b.progress.remainingDays
-      : Infinity;
-    if (remainingA !== remainingB) {
-      return remainingA - remainingB;
-    }
-    const payoutA = Number.isFinite(a?.payout) ? a.payout : 0;
-    const payoutB = Number.isFinite(b?.payout) ? b.payout : 0;
-    if (payoutA !== payoutB) {
-      return payoutB - payoutA;
-    }
-    const orderA = Number.isFinite(a?.orderIndex) ? a.orderIndex : 0;
-    const orderB = Number.isFinite(b?.orderIndex) ? b.orderIndex : 0;
-    return orderA - orderB;
-  });
-}
-
-registerFocusBucket({
-  name: 'commitment',
-  comparator: sortCommitmentEntries,
-  modes: {
-    money: { order: ['commitment', 'hustle', 'upgrade'] },
-    upgrades: { order: ['commitment', 'upgrade', 'hustle'] },
-    balanced: { order: ['commitment', 'upgrade', 'hustle'], interleave: ['commitment', 'upgrade', 'hustle'] }
-  }
-});
-registerFocusBucket({ name: 'hustle', comparator: sortHustleEntries });
-registerFocusBucket({ name: 'upgrade', comparator: sortUpgradeEntries });
-
-function interleaveEntries(first = [], second = []) {
-  const results = [];
-  const max = Math.max(first.length, second.length);
-  for (let index = 0; index < max; index += 1) {
-    if (first[index]) {
-      results.push(first[index]);
-    }
-    if (second[index]) {
-      results.push(second[index]);
-    }
-  }
-  return results;
-}
-
-function resolveProgressStep(entry = {}) {
-  const progress = entry?.progress || {};
-  const rawStep = Number(progress.stepHours);
-  if (Number.isFinite(rawStep) && rawStep > 0) {
-    const remaining = Number(progress.hoursRemaining);
-    if (Number.isFinite(remaining) && remaining >= 0) {
-      return Math.max(0, Math.min(rawStep, remaining));
-    }
-    return rawStep;
-  }
-
-  const duration = Number(entry?.durationHours);
-  if (Number.isFinite(duration) && duration > 0) {
-    return duration;
-  }
-  return 0;
-}
-
-function createProgressHandler(entry = {}) {
-  const progress = entry?.progress || {};
-  const definitionId = progress.definitionId || entry?.definitionId;
-  const instanceId = progress.instanceId || entry?.instanceId;
-  if (!definitionId || !instanceId) {
-    return null;
-  }
-
-  return () => {
-    const definition = getActionDefinition(definitionId);
-    if (!definition) {
-      return { success: false };
-    }
-
-    const metadata = progress.metadata;
-    const remaining = Number(progress.hoursRemaining);
-    const step = resolveProgressStep(entry);
-
-    const requiresManualCompletion = progress.completion === 'manual';
-    const hasRemaining = Number.isFinite(remaining) ? remaining > 0 : true;
-    const amount = step > 0 ? step : Math.max(0, remaining);
-
-    if (requiresManualCompletion && (!hasRemaining || amount <= 0)) {
-      completeActionInstance(definition, { id: instanceId }, { metadata });
-      return { success: true, hours: 0, completed: true };
-    }
-
-    const hours = amount > 0 ? amount : step;
-    if (!Number.isFinite(hours) || hours <= 0) {
-      completeActionInstance(definition, { id: instanceId }, { metadata });
-      return { success: true, hours: 0, completed: true };
-    }
-
-    const state = getState();
-    const available = Number(state?.timeLeft);
-    if (Number.isFinite(available) && available < hours) {
-      if (typeof addLog === 'function') {
-        addLog(
-          `You need ${formatHours(hours)} focus free before logging that commitment. Wrap another task first or rest up.`,
-          'warning'
-        );
-      }
-      return { success: false };
-    }
-
-    spendTime(hours);
-    advanceActionInstance(definition, { id: instanceId }, { hours, metadata });
-
-    const isStudy =
-      definition.studyTrackId ||
-      definition?.progress?.type === 'study' ||
-      progress?.type === 'study';
-    if (isStudy) {
-      const resolvedTrackId = (() => {
-        const directIdCandidates = [
-          definition.studyTrackId,
-          progress?.studyTrackId,
-          progress?.trackId,
-          definition?.progress?.studyTrackId,
-          definition?.progress?.trackId
-        ].filter(id => typeof id === 'string' && id.trim().length > 0);
-        if (directIdCandidates.length > 0) {
-          return directIdCandidates[0];
-        }
-        const stripStudyPrefix = value =>
-          typeof value === 'string' && value.startsWith('study-') ? value.slice('study-'.length) : null;
-        return (
-          stripStudyPrefix(definition?.id) ||
-          stripStudyPrefix(progress?.definitionId) ||
-          null
-        );
-      })();
-
-      if (resolvedTrackId) {
-        allocateDailyStudy({ trackIds: [resolvedTrackId] });
-      } else {
-        allocateDailyStudy();
-      }
-    }
-    checkDayEnd();
-
-    return { success: true, hours };
-  };
-}
-
-function attachProgressHandlers(entries = []) {
-  return entries.map(entry => {
-    if (!entry?.progress) {
-      return entry;
-    }
-    const handler = createProgressHandler(entry);
-    if (handler) {
-      entry.onClick = handler;
-    }
-    return entry;
-  });
-}
-
-function getBucketName(entry) {
-  if (!entry || typeof entry !== 'object') {
-    return DEFAULT_FOCUS_BUCKET;
-  }
-  const explicitBucket = typeof entry.focusBucket === 'string' ? entry.focusBucket : null;
-  if (explicitBucket) {
-    return explicitBucket;
-  }
-  const categoryBucket = typeof entry.focusCategory === 'string' ? entry.focusCategory : null;
-  if (categoryBucket) {
-    return categoryBucket;
-  }
-  return DEFAULT_FOCUS_BUCKET;
-}
-
-function collectBuckets(entries = []) {
-  const buckets = new Map();
-  entries.filter(Boolean).forEach(entry => {
-    const bucketName = getBucketName(entry);
-    if (!buckets.has(bucketName)) {
-      buckets.set(bucketName, []);
-    }
-    buckets.get(bucketName).push(entry);
-  });
-  return buckets;
-}
-
-function sortBuckets(bucketMap = new Map()) {
-  const sorted = new Map();
-  bucketMap.forEach((bucketEntries, bucketName) => {
-    const comparator = getFocusBucketComparator(bucketName);
-    sorted.set(bucketName, comparator(bucketEntries));
-  });
-  return sorted;
-}
-
-export function applyFocusOrdering(entries = [], mode = 'balanced') {
-  if (!Array.isArray(entries) || entries.length === 0) {
-    return entries;
-  }
-
-  const buckets = collectBuckets(entries);
-  if (buckets.size === 0) {
-    return entries;
-  }
-
-  const sortedBuckets = sortBuckets(buckets);
-  const resolvedModeConfig = (() => {
-    const direct = getFocusModeConfig(mode);
-    if (direct) {
-      return direct;
-    }
-    if (mode !== 'balanced') {
-      const fallback = getFocusModeConfig('balanced');
-      if (fallback) {
-        return fallback;
-      }
-    }
-    return { order: [], interleave: [] };
-  })();
-  const order = Array.isArray(resolvedModeConfig?.order) ? resolvedModeConfig.order : [];
-  const interleaveBuckets = Array.isArray(resolvedModeConfig?.interleave)
-    ? resolvedModeConfig.interleave
-    : [];
-
-  const results = [];
-  const usedEntries = new Set();
-
-  const addEntry = entry => {
-    if (!entry || usedEntries.has(entry)) {
-      return;
-    }
-    usedEntries.add(entry);
-    results.push(entry);
-  };
-
-  const processedBuckets = new Set();
-
-  if (interleaveBuckets.length > 0) {
-    let interleaved = [];
-    interleaveBuckets.forEach((bucketName, index) => {
-      const bucketEntries = sortedBuckets.get(bucketName) || [];
-      processedBuckets.add(bucketName);
-      if (index === 0) {
-        interleaved = [...bucketEntries];
-      } else {
-        interleaved = interleaveEntries(interleaved, bucketEntries);
-      }
-    });
-    interleaved.forEach(addEntry);
-  }
-
-  order.forEach(bucketName => {
-    processedBuckets.add(bucketName);
-    const bucketEntries = sortedBuckets.get(bucketName) || [];
-    bucketEntries.forEach(addEntry);
-  });
-
-  sortedBuckets.forEach((bucketEntries, bucketName) => {
-    if (processedBuckets.has(bucketName)) {
-      return;
-    }
-    bucketEntries.forEach(addEntry);
-  });
-
-  return results;
-}
 
 function applyImmediateTimeDelta(model, hours) {
   if (!model || !Number.isFinite(hours) || hours <= 0) return;
@@ -539,10 +221,11 @@ export function render(model = {}) {
   todoState.setPendingEntries(orderedPending);
 
   todoDom.renderHours(elements, viewModel, formatHours);
-  todoDom.updateNote(elements?.note, viewModel, orderedPending.length);
+  const emptyMessage = viewModel.emptyMessage || DEFAULT_TODO_EMPTY_MESSAGE;
+  todoDom.updateNote(elements?.note, { ...viewModel, emptyMessage }, orderedPending.length);
 
   if (!orderedPending.length) {
-    todoDom.renderEmptyState(elements?.list, viewModel.emptyMessage, callEndDay);
+    todoDom.renderEmptyState(elements?.list, emptyMessage, callEndDay);
   } else {
     todoDom.renderPending(elements?.list, orderedPending, viewModel, handleCompletion);
   }
